Refactored test structure
