+++ tests/Proto.Actor.Tests/MiddlewareTests.cs
@@ -15,7 +15,7 @@ namespace Proto.Tests
         public void Given_Middleware_Should_Call_In_Order_Then_Actor()
         {
             var logs = new List<string>();
-            var testMailbox = new ActorFixture.TestMailbox();
+            var testMailbox = new Fixture.TestMailbox();
             var props = Actor.FromFunc(c =>
                 {
                     if (c.Message is string)
